+++ ictools/hipchat.py
@@ -1,13 +1,12 @@
+import argparse
 import logging
-import os
-import pathlib
 import sys
 import urllib.parse
 
 import maya
 import requests
 
-from ictools import auth, cli, io
+from ictools import auth, cli, io, version
 
 
 _logger = logging.getLogger(__name__)
@@ -25,16 +24,17 @@ class Connection(object):
         response.raise_for_status()
 
     def fetch_messages(self, room, start, end):
-        logging.getLogger('requests').setLevel(logging.DEBUG)
+        start = start.datetime().isoformat()
+        end = end.datetime().isoformat()
         self.logger.info('fetching messages from %s between %s and %s',
-                         room, start.isoformat(), end.isoformat())
+                         room, start, end)
         response = self.session.get(
             'https://api.hipchat.com/v2/room/{}/history'.format(
                 urllib.parse.quote(room)), auth=self.auth,
-            params={'date': end.isoformat(), 'end-date': start.isoformat()})
+            params={'date': end, 'end-date': start})
         if response.status_code == 404:
             self.logger.error('room %s not found', room)
-            return None
+            return []
         response.raise_for_status()
 
         return _augment_messages(response.json()['items'], room)
@@ -49,23 +49,26 @@ def _augment_messages(messages, room):
 
 
 def scan_room():
-    logger = _logger.getChild('scan_room')
     cli.configure_logging()
+    logger = _logger.getChild('scan_room')
     api_token = cli.require_environment(logger, 'HIPCHAT_TOKEN')
 
-    if len(sys.argv) < 3:
-        prog_name = pathlib.Path(sys.argv[0]).stem
-        sys.stderr.write(
-            'Usage: {} START END ROOM [ROOM...]\n'.format(prog_name))
-        sys.exit(os.EX_USAGE)
+    parser = argparse.ArgumentParser(
+        description='Retrieve messages from HipChat')
+    parser.add_argument('--version', action='version', version=version)
+    parser.add_argument('start_date', metavar='START', type=maya.parse,
+                        help='timestamp to start retrieving messages from')
+    parser.add_argument('end_date', metavar='END', type=maya.parse,
+                        help='timestamp to stop retrieving messages at')
+    parser.add_argument('rooms', metavar='ROOM', nargs='+',
+                        help='one or more rooms to retrieve messages from')
+    args = parser.parse_args()
 
     conn = Connection(api_token)
-    earliest = maya.parse(sys.argv[1]).datetime()
-    latest = maya.parse(sys.argv[2]).datetime()
     messages = []
-    for room in sys.argv[3:]:
-        items = conn.fetch_messages(room, earliest, latest)
-        logger.debug('found %d messages in %d', len(items), room)
+    for room in args.rooms:
+        items = conn.fetch_messages(room, args.start_date, args.end_date)
+        logger.debug('found %d messages in %s', len(items), room)
         messages.extend(items)
 
     logger.info('found %d messages', len(messages))