+++ docs/history.rst
@@ -1,6 +1,12 @@
 Release History
 ===============
 
+`Next Release`_
+---------------
+* added "metadata" element to all JSON output.
+* switched to using :mod:`argparse` for parsing.  The only observable effect
+  of this is that ``--help`` and ``--version`` are now supported.
+
 `0.1.0`_ (23-Jan-2017)
 ----------------------
 * add :ref:`scan_hipchat_room` utility