+++ docs/history.rst
@@ -1,8 +1,8 @@
 Release History
 ===============
 
-`Next Release`_
----------------
+`0.2.0`_ (11-Feb-2017)
+----------------------
 * added "metadata" element to all JSON output.
 * switched to using :mod:`argparse` for parsing.  The only observable effect
   of this is that ``--help`` and ``--version`` are now supported.
@@ -16,5 +16,6 @@ Release History
 * added :ref:`scan_hipchat_room` utility
 * added :ref:`list_pagerduty_incidents` utility
 
-.. _Next Release: https://github.com/dave-shawley/ictools/compare/0.1.0...HEAD
+.. _Next Release: https://github.com/dave-shawley/ictools/compare/0.2.0...HEAD
+.. _0.2.0: https://github.com/dave-shawley/ictools/compare/0.1.0...0.2.0
 .. _0.1.0: https://github.com/dave-shawley/ictools/compare/0.0.0...0.1.0